On peut maintenant parier sur un match : le pari s'enregistre, les credits descendent et les cotes s'ajustent. Fait : controle de doublon de paris (pas fait : message au debut de bet.php pour dire que ca va annuler le dernier pari et rappel du pari) ; controle sur les pauvres qui misent trop (pas fait : message qui dit que t'es trop pauvre). Reste a faire : un bouton 'modifier pari', un bouton 'fin du pari', un champ dans la db 'resultat final' qui va dire qui gagne quand on appui sur le bouton decris juste precedemment, un form dans profil.php qui permet de modifier ses coordonnees et notamment son adresse, l'affichage de ses coordonnees dans profil.php et notamment son adresse, et puis tout ce qui est decrit sur le sujet.pdf en fait xD

+++ db_install.php
@@ -25,9 +25,9 @@ $qTbMatch = "CREATE TABLE IF NOT EXISTS `match` (
   `team1` varchar(25) NOT NULL,
   `team2` varchar(25) NOT NULL,
   `date` date NOT NULL,
-  `odds1` int(3) NOT NULL,  /*il faut mettre un float */
-  `odds2` int(3) NOT NULL,  /*il faut mettre un float */
-  `draw` int(3) NOT NULL,   /*il faut mettre un float */
+  `odds1` float NOT NULL,  
+  `odds2` float NOT NULL,  
+  `draw` float NOT NULL,   
   PRIMARY KEY (`ref`)
 ) ENGINE=InnoDB;";
